Move new async to asynch module to guard for models without flash interrupt
