+++ embassy-stm32/src/flash/mod.rs
@@ -1,10 +1,12 @@
 use embedded_storage::nor_flash::{NorFlashError, NorFlashErrorKind};
 
 #[cfg(all(feature = "nightly", flash_f4))]
-pub mod asynch;
+mod asynch;
 #[cfg(flash)]
 mod common;
 
+#[cfg(all(feature = "nightly", flash_f4))]
+pub use asynch::InterruptHandler;
 #[cfg(flash)]
 pub use common::*;
